move crawler instance to crawl.py
