+++ app/views.py
@@ -33,7 +33,8 @@ from .models import Crawl, DataSource, Dashboard, Plot, Project
 from .forms import CrawlForm, MonitorDataForm, PlotForm, ContactForm, \
                     DashboardForm, ProjectForm
 from .mail import send_email
-from .config import ADMINS, DEFAULT_MAIL_SENDER, CRAWLER_PATH
+from .config import ADMINS, DEFAULT_MAIL_SENDER
+from .crawls import CrawlAcheInstance
 from .auth import requires_auth
 from .plotting import plot_builder
 
@@ -103,35 +104,6 @@ def add_project():
 # -----------------------------------------------------------------------------
 
 
-class CrawlInstance(object):
-
-    def __init__(self, seeds_list, model_name):
-        self.seeds_list = seeds_list
-        self.model_name = model_name
-        self.proc = None
-
-    def start(self):
-        self.proc = subprocess.Popen('./run_crawler.sh {0} conf/ conf/seeds/{1} conf/models/{2}/'
-                                     .format(CRAWLER_PATH, self.seeds_list, self.model_name), shell=True)
-        #self.proc = subprocess.Popen('./count_things.sh', shell=True)
-        return self.proc.pid
-
-    def stop(self):
-        if self.proc is not None:
-            print("Killing %s" % str(self.proc.pid))
-            self.proc.kill()
-            proc2 = subprocess.Popen('./stop_crawler.sh {0}'.format((CRAWLER_PATH)), shell=True)
-
-    def status(self):
-        if self.proc is None:
-            return "No process exists"
-        elif self.proc.returncode is None:
-            return "Running"
-        elif self.proc.returncode < 0:
-            return "Stopped (Unused)"
-        else:
-            return "An error occurred"
-
 
 @app.route('/<project_name>/add_crawl', methods=['GET', 'POST'])
 def add_crawl(project_name):
@@ -193,7 +165,7 @@ def run_crawl(project_name, crawl_name):
         crawl = Crawl.query.filter_by(name=crawl_name).first()
         seeds_list = crawl.seeds_list
         model_name = crawl.data_model
-        crawl_instance = CrawlInstance(seeds_list, model_name)
+        crawl_instance = CrawlAcheInstance(seeds_list, model_name)
         pid = crawl_instance.start()
         CRAWLS_RUNNING[key] = crawl_instance
         return "Crawl running"
